refactor: RestartNumber 제거 및 RestartState 1개로 운용
